pyproject.toml
[tool.poetry]
name = "goktugv2"
version = "0.1.0"
description = ""
authors = ["Apollo314 <abidinsefa@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
pyppeteer = "^1.0.2"
uvicorn = "^0.18.2"
uritemplate = "^4.1.1"
python-dotenv = "^0.20.0"
django-simple-history = "^3.1.1"
Pillow = "^9.2.0"
drf-spectacular = "^0.24.2"
django-rest-knox = "^4.2.0"
psycopg2-binary = "^2.9.5"
gunicorn = "^20.1.0"
orjson = "^3.8.1"
pydantic = "^1.10.2"
drf-extra-fields = "^3.4.1"
django-extensions = "^3.2.1"
drf-standardized-errors = "^0.12.3"
lxml = "^4.9.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
PyYAML = "^6.0"
coverage = "^6.4.2"
ipython = "^8.5.0"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}
ipykernel = "^6.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
